requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn[standard]>=0.30.0
asyncpg==0.29.0
python-dotenv>=0.19.0,<1.0.0
sqlalchemy==2.0.30
psycopg2==2.9.9
PyJWT==2.8.0
passlib==1.7.4
python-dotenv==0.21.1
sqlalchemy==2.0.30